+++ sdk/identity/azure_identity/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ### Breaking Changes
 
+- `ClientCertificateCredential::new()` takes `Option<ClientCertificateCredentialOptions>` instead of `impl Into<ClientCertificateCredentialOptions>`
+- Credential constructors return an error when given a non-HTTPS authority host.
+- Replaced credential-specific `authority_host` options with `azure_core::cloud::CloudConfiguration` configured via `ClientOptions.cloud`.
+
 ### Bugs Fixed
 
 ### Other Changes